requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-socketio>=5.3.0
eventlet>=0.35.0
pyyaml>=6.0.1
gunicorn>=21.2.0
python-dotenv>=1.0.0
